requirements.txt
vosk
sounddevice
pillow


numpy
